docs(Dropdown): render Dropdown example stories live instead of StackBlitz
The stories under Components/Dropdown/With Select, With Button and Link, and
With AutoComplete embedded StackBlitz sandboxes, which need cross-origin
isolation to boot their WebContainer. Storybook does not serve the COOP/COEP
headers for that, so the examples showed a blank or perpetually loading canvas
instead of a Dropdown.

Render the examples as real components so they behave like the rest of the
stories, and drop the code strings that existed only to feed those editors. The
snippets the docs page still renders through Sandpack stay.

Several of the ported code strings could never have run, since nothing
type-checked them:

- WithTooltipStory opened a DropdownIconButton and closed it with
  </DropdownLink>, and had an unbalanced onOpenChange callback.
- WithLinkStory had the same unbalanced onOpenChange.
- clearOnDismiss and withErrorState passed an onDismiss prop that Dropdown does
  not have; they now use onOpenChange.
- WithHeaderFooterScroll reused the same ActionListItem value across items,
  which confuses selection identity.
- responsiveBottomSheet depended on a hardcoded isMobile constant the reader had
  to edit; it now uses useIsMobile so the BottomSheet swap is actually
  responsive.
- The tag row in WithControlledMultiSelect stretched its tags to 32px instead of
  their natural 21px, because the flex row set a minHeight without alignItems.

Also fix the sidebar CSS that hides internal stories: it matched the story
anchor rather than the row wrapper, so every hidden story left an empty 28px row
behind in Storybook 10.
