+++ .changeset/dropdown-examples-live-preview.md
@@ -1,0 +1,5 @@
+---
+'@razorpay/blade': patch
+---
+
+docs(Dropdown): render Dropdown example stories live instead of StackBlitz sandboxes
