+++ packages/blade/src/components/Dropdown/docs/DropdownWithButton.stories.tsx
@@ -3,18 +3,13 @@ import { DropdownButton } from '../DropdownButton';
 import { Dropdown, DropdownLink, DropdownOverlay } from '..';
 import { DropdownFooter, DropdownHeader } from '../DropdownHeaderFooter';
 import { DropdownIconButton } from '../DropdownIconButton';
-import {
-  WithControlledMenuStory,
-  WithControlledMultiSelectMenuStory,
-  WithLinkStory,
-  WithAutoPositioningMenuStory,
-  WithSimpleMenuStory,
-  WithIconButtonStory,
-  WithTooltipStory,
-} from './stories';
-import { Sandbox } from '~utils/storybook/Sandbox';
 import { Box } from '~components/Box';
-import { ActionList, ActionListItem, ActionListItemIcon } from '~components/ActionList';
+import {
+  ActionList,
+  ActionListItem,
+  ActionListItemIcon,
+  ActionListSection,
+} from '~components/ActionList';
 import {
   BoxIcon,
   CheckIcon,
@@ -22,6 +17,7 @@ import {
   ChevronUpIcon,
   ClockIcon,
   CloseIcon,
+  MyAccountIcon,
   StarIcon,
 } from '~components/Icons';
 import { Text } from '~components/Typography';
@@ -29,6 +25,7 @@ import { Checkbox } from '~components/Checkbox';
 import { Button } from '~components/Button';
 import { Badge } from '~components/Badge';
 import { Amount } from '~components/Amount';
+import { Tag } from '~components/Tag';
 import { AutoComplete } from '~components/Input/DropdownInputTriggers';
 import { Tooltip, TooltipInteractiveWrapper } from '~components/Tooltip';
 
@@ -53,61 +50,355 @@ const DropdownStoryMeta = {
 
 export const Default = (): React.ReactElement => {
   return (
-    <Sandbox
-      padding="spacing.0"
-      editorHeight="100vh"
-      uri="https://blade.razorpay.com/iframe.html?id=components-dropdown-with-button--default&args=&viewMode=story"
-    >
-      {WithSimpleMenuStory}
-    </Sandbox>
+    <Box minHeight="200px" width={{ base: '100%', m: '500px' }} padding="spacing.5">
+      <Dropdown>
+        <DropdownButton icon={MyAccountIcon} variant="secondary">
+          My Account
+        </DropdownButton>
+        <DropdownOverlay>
+          <ActionList>
+            <ActionListSection title="Account @saurabh">
+              <ActionListItem
+                title="My Profile"
+                value="profile"
+                href="https://youtu.be/4qRZmFYdozY?t=33"
+                target="_blank"
+              />
+              <ActionListItem
+                title="Dashboard"
+                value="dashboard"
+                href="https://dashboard.razorpay.com/"
+              />
+              <ActionListItem
+                title="Settings"
+                value="settings"
+                href="https://memezila.com/Me-changing-the-phone-language-just-for-fun-Couldnt-find-language-setting-now-meme-5150"
+              />
+            </ActionListSection>
+            <ActionListItem
+              intent="negative"
+              title="Log Out"
+              value="logout"
+              onClick={() => {
+                console.log('Logging out');
+              }}
+            />
+          </ActionList>
+        </DropdownOverlay>
+      </Dropdown>
+    </Box>
   );
 };
 
 export const WithLink = (): React.ReactElement => {
+  const [status, setStatus] = React.useState<string | undefined>('latest-added');
+  const [isDropdownOpen, setIsDropdownOpen] = React.useState(false);
+
   return (
-    <Sandbox padding="spacing.0" editorHeight="100vh">
-      {WithLinkStory}
-    </Sandbox>
+    <Box padding="spacing.10" display="flex" alignItems="center" gap="spacing.2">
+      <Text>Sort By</Text>
+      <Box flex="1">
+        <Dropdown isOpen={isDropdownOpen} onOpenChange={setIsDropdownOpen}>
+          <DropdownLink
+            icon={isDropdownOpen ? ChevronUpIcon : ChevronDownIcon}
+            iconPosition="right"
+          >
+            {status ?? ''}
+          </DropdownLink>
+          <DropdownOverlay>
+            <ActionList>
+              <ActionListItem
+                onClick={({ name, value }) => {
+                  console.log({ name, value });
+                  setStatus(name);
+                }}
+                isSelected={status === 'latest-added'}
+                title="Latest Added"
+                value="latest-added"
+              />
+              <ActionListItem
+                onClick={({ name, value }) => {
+                  console.log({ name, value });
+                  setStatus(name);
+                }}
+                isSelected={status === 'latest-invoice'}
+                title="Latest Invoice"
+                value="latest-invoice"
+              />
+              <ActionListItem
+                onClick={({ name, value }) => {
+                  console.log({ name, value });
+                  setStatus(name);
+                }}
+                isSelected={status === 'oldest-due-date'}
+                title="Oldest Due Date"
+                value="oldest-due-date"
+              />
+            </ActionList>
+          </DropdownOverlay>
+        </Dropdown>
+      </Box>
+    </Box>
   );
 };
 
 export const WithIconButton = (): React.ReactElement => {
+  const [status, setStatus] = React.useState<string | undefined>('latest-added');
+
   return (
-    <Sandbox padding="spacing.0" editorHeight="100vh">
-      {WithIconButtonStory}
-    </Sandbox>
+    <Box padding="spacing.10">
+      <Dropdown>
+        <DropdownIconButton icon={BoxIcon} accessibilityLabel="Set Status" />
+        <DropdownOverlay>
+          <ActionList>
+            <ActionListItem
+              onClick={({ name, value }) => {
+                console.log({ name, value });
+                setStatus(name);
+              }}
+              isSelected={status === 'latest-added'}
+              title="Latest Added"
+              value="latest-added"
+            />
+            <ActionListItem
+              onClick={({ name, value }) => {
+                console.log({ name, value });
+                setStatus(name);
+              }}
+              isSelected={status === 'latest-invoice'}
+              title="Latest Invoice"
+              value="latest-invoice"
+            />
+            <ActionListItem
+              onClick={({ name, value }) => {
+                console.log({ name, value });
+                setStatus(name);
+              }}
+              isSelected={status === 'oldest-due-date'}
+              title="Oldest Due Date"
+              value="oldest-due-date"
+            />
+          </ActionList>
+        </DropdownOverlay>
+      </Dropdown>
+    </Box>
   );
 };
 
 export const WithAutoPositioning = (): React.ReactElement => {
   return (
-    <Sandbox padding="spacing.0" editorHeight="100vh">
-      {WithAutoPositioningMenuStory}
-    </Sandbox>
+    <Box>
+      <Box display="inline-flex" position="fixed" left="spacing.5" top="spacing.5">
+        <Dropdown>
+          <DropdownButton>Top Left Menu</DropdownButton>
+          <DropdownOverlay>
+            <ActionList>
+              <ActionListItem title="Apples" value="Apples" />
+              <ActionListItem title="Appricots" value="Appricots" />
+            </ActionList>
+          </DropdownOverlay>
+        </Dropdown>
+      </Box>
+      <Box display="inline-flex" position="fixed" right="spacing.5" top="spacing.5">
+        <Dropdown>
+          <DropdownButton>Top Right Menu</DropdownButton>
+          <DropdownOverlay>
+            <ActionList>
+              <ActionListItem title="Apples" value="Apples" />
+              <ActionListItem title="Appricots" value="Appricots" />
+            </ActionList>
+          </DropdownOverlay>
+        </Dropdown>
+      </Box>
+      <Box display="inline-flex" position="fixed" right="spacing.5" bottom="spacing.5">
+        <Dropdown>
+          <DropdownButton>Bottom Right Menu</DropdownButton>
+          <DropdownOverlay>
+            <ActionList>
+              <ActionListItem title="Apples" value="Apples" />
+              <ActionListItem title="Appricots" value="Appricots" />
+            </ActionList>
+          </DropdownOverlay>
+        </Dropdown>
+      </Box>
+      <Box display="inline-flex" position="fixed" left="spacing.5" bottom="spacing.5">
+        <Dropdown>
+          <DropdownButton>Bottom Left Menu</DropdownButton>
+          <DropdownOverlay>
+            <ActionList>
+              <ActionListItem title="Apples" value="Apples" />
+              <ActionListItem title="Appricots" value="Appricots" />
+            </ActionList>
+          </DropdownOverlay>
+        </Dropdown>
+      </Box>
+    </Box>
   );
 };
 
 export const WithControlledMenu = (): React.ReactElement => {
+  const [status, setStatus] = React.useState<string | undefined>();
+
   return (
-    <Sandbox padding="spacing.0" editorHeight="100vh">
-      {WithControlledMenuStory}
-    </Sandbox>
+    <Box minHeight="200px" padding="spacing.5">
+      <Dropdown>
+        <DropdownButton variant="tertiary">Status: {status ?? ''}</DropdownButton>
+        <DropdownOverlay>
+          <ActionList>
+            <ActionListItem
+              onClick={({ name, value }) => {
+                console.log({ name, value });
+                setStatus(name);
+              }}
+              leading={<ActionListItemIcon icon={CheckIcon} />}
+              isSelected={status === 'approve'}
+              title="Approve"
+              value="approve"
+            />
+            <ActionListItem
+              onClick={({ name, value }) => {
+                console.log({ name, value });
+                setStatus(name);
+              }}
+              leading={<ActionListItemIcon icon={ClockIcon} />}
+              isSelected={status === 'in-progress'}
+              title="In Progress"
+              value="in-progress"
+            />
+            <ActionListItem
+              onClick={({ name, value }) => {
+                console.log({ name, value });
+                setStatus(name);
+              }}
+              leading={<ActionListItemIcon icon={CloseIcon} />}
+              isSelected={status === 'reject'}
+              title="Reject"
+              value="reject"
+              intent="negative"
+            />
+          </ActionList>
+        </DropdownOverlay>
+      </Dropdown>
+    </Box>
   );
 };
 
 export const WithControlledMultiSelect = (): React.ReactElement => {
+  const [filters, setFilters] = React.useState<string[]>([]);
+
+  const toggleSelection = ({ name, value }: { name: string; value?: boolean }): void => {
+    if (value) {
+      // Value is true which means it is selected. Then we deselect it.
+      const existingItemIndex = filters.indexOf(name);
+      setFilters([...filters.slice(0, existingItemIndex), ...filters.slice(existingItemIndex + 1)]);
+    } else {
+      setFilters([...filters, name]);
+    }
+  };
+
   return (
-    <Sandbox padding="spacing.0" editorHeight="100vh">
-      {WithControlledMultiSelectMenuStory}
-    </Sandbox>
+    <Box minHeight="200px" padding="spacing.5">
+      <Box
+        display="flex"
+        alignItems="center"
+        flexWrap="wrap"
+        paddingBottom="spacing.5"
+        minHeight="spacing.10"
+      >
+        {filters.map((filter) => (
+          <Tag
+            key={filter}
+            marginRight="spacing.3"
+            onDismiss={() => {
+              toggleSelection({ name: filter, value: true });
+            }}
+          >
+            {filter}
+          </Tag>
+        ))}
+      </Box>
+      <Dropdown selectionType="multiple">
+        <DropdownButton variant="tertiary">Filters: {filters.length} Applied</DropdownButton>
+        <DropdownOverlay>
+          <ActionList>
+            <ActionListItem
+              onClick={({ name, value }) => {
+                console.log({ name, value });
+                toggleSelection({ name, value });
+              }}
+              isSelected={filters.includes('< 3 months')}
+              title="Last 3 months"
+              value="< 3 months"
+            />
+            <ActionListItem
+              onClick={({ name, value }) => {
+                console.log({ name, value });
+                toggleSelection({ name, value });
+              }}
+              isSelected={filters.includes('> 1000rs')}
+              title="More than 1000rs"
+              value="> 1000rs"
+            />
+            <ActionListItem
+              onClick={({ name, value }) => {
+                console.log({ name, value });
+                toggleSelection({ name, value });
+              }}
+              isSelected={filters.includes('failed')}
+              title="Failed Transactions"
+              value="failed"
+            />
+          </ActionList>
+        </DropdownOverlay>
+      </Dropdown>
+    </Box>
   );
 };
 
 export const WithTooltip = (): React.ReactElement => {
+  const [status, setStatus] = React.useState<string | undefined>('latest-added');
+
   return (
-    <Sandbox padding="spacing.0" editorHeight="100vh">
-      {WithTooltipStory}
-    </Sandbox>
+    <Box padding="spacing.10">
+      <Tooltip content="Change Status">
+        <TooltipInteractiveWrapper>
+          <Dropdown>
+            <DropdownIconButton icon={BoxIcon} accessibilityLabel="Set Status" />
+            <DropdownOverlay>
+              <ActionList>
+                <ActionListItem
+                  onClick={({ name, value }) => {
+                    console.log({ name, value });
+                    setStatus(name);
+                  }}
+                  isSelected={status === 'latest-added'}
+                  title="Latest Added"
+                  value="latest-added"
+                />
+                <ActionListItem
+                  onClick={({ name, value }) => {
+                    console.log({ name, value });
+                    setStatus(name);
+                  }}
+                  isSelected={status === 'latest-invoice'}
+                  title="Latest Invoice"
+                  value="latest-invoice"
+                />
+                <ActionListItem
+                  onClick={({ name, value }) => {
+                    console.log({ name, value });
+                    setStatus(name);
+                  }}
+                  isSelected={status === 'oldest-due-date'}
+                  title="Oldest Due Date"
+                  value="oldest-due-date"
+                />
+              </ActionList>
+            </DropdownOverlay>
+          </Dropdown>
+        </TooltipInteractiveWrapper>
+      </Tooltip>
+    </Box>
   );
 };
